Extract shared test infrastructure into common module
Move RT, APP, app(), TestUser, WithAuth, and helpers into
tests/common/mod.rs so new test files can reuse them without
duplicating boilerplate.
