+++ passwords/api/tests/integration_tests.rs
@@ -4,19 +4,6 @@
 //! MongoDB instance (configured via `.env`). They use Axum's in-process
 //! `Router` with `tower::ServiceExt::oneshot` so no actual TCP port is opened.
 //!
-//! ## Architecture
-//!
-//! - **Shared runtime**: A single `LazyLock<Runtime>` keeps the MongoDB
-//!   connection pool alive across all tests. Each `#[test]` function calls
-//!   `run()` which delegates to `RT.block_on()`.
-//!
-//! - **Shared router**: A single `LazyLock<Router>` is initialized once on the
-//!   shared runtime. Each request clones it (Axum routers are cheaply cloneable).
-//!
-//! - **TestUser RAII**: Each test that needs a user creates a `TestUser` whose
-//!   `Drop` impl deletes it from the database. A separate OS thread is used
-//!   for the cleanup to avoid nesting `block_on` calls.
-//!
 //! ## Running
 //!
 //! ```sh
@@ -24,135 +11,16 @@
 //! cargo test --test integration_tests --features test-helpers
 //! ```
 
+mod common;
+
 use axum::body::Body;
-use axum::{Router, middleware::from_fn, extract::ConnectInfo};
+use axum::{middleware::from_fn, extract::ConnectInfo};
+use common::{app, body_string, parse_json, run, TestUser, WithAuth};
 use http::{Request, StatusCode};
-use mongodb::bson::oid::ObjectId;
-use passwords::{build_router, build_router_with_burst};
-use passwords::db;
+use passwords::build_router;
 use std::net::SocketAddr;
-use std::sync::LazyLock;
 use std::time::Duration;
 use tower::ServiceExt;
-
-const TEST_PW: &str = "test_password_abc123";
-
-// ---------------------------------------------------------------------------
-// Single shared runtime – keeps the MongoDB connection pool alive across tests.
-// Shared Axum router + DB connection (initialized once on the shared runtime).
-// ---------------------------------------------------------------------------
-
-static RT: LazyLock<tokio::runtime::Runtime> = LazyLock::new(|| {
-    tokio::runtime::Builder::new_multi_thread()
-        .enable_all()
-        .build()
-        .expect("Failed to create shared tokio runtime")
-});
-
-static APP: LazyLock<Router> = LazyLock::new(|| {
-    RT.block_on(async {
-        dotenv::dotenv().ok();
-        db::connect().await.expect("Failed to connect to test DB");
-        // use a very large burst so ordinary tests aren't disrupted by our
-        // rate limiter; stress test will create its own router below.
-        build_router_with_burst(1_000_000)
-    })
-});
-
-/// Returns a fresh clone of the shared router (needed because `oneshot` consumes the service).
-fn app() -> Router {
-    // For tests we never run a real TCP server, so the GovernorLayer's default
-    // PeerIpKeyExtractor would fail to extract a peer IP (leading to 500
-    // errors).  Inject a dummy ConnectInfo using middleware so the rate limiter
-    // sees a valid address on every request.
-    let addr = SocketAddr::from(([127, 0, 0, 1], 0));
-    APP.clone().layer(from_fn(move |mut req: Request<Body>, next: axum::middleware::Next| async move {
-        req.extensions_mut().insert(ConnectInfo(addr));
-        next.run(req).await
-    }))
-}
-
-// ---------------------------------------------------------------------------
-// WithAuth: extension trait for attaching credentials to test requests.
-// ---------------------------------------------------------------------------
-
-trait WithAuth {
-    fn auth(self, user: &str, pw: &str) -> Self;
-}
-
-impl WithAuth for http::request::Builder {
-    fn auth(self, user: &str, pw: &str) -> Self {
-        self.header("x-username", user)
-            .header("x-password", pw)
-    }
-}
-
-// ---------------------------------------------------------------------------
-// TestUser: RAII guard that generates a unique username and deletes it on drop.
-// ---------------------------------------------------------------------------
-
-struct TestUser {
-    username: String,
-    password: String,
-}
-
-impl TestUser {
-    fn new() -> Self {
-        Self {
-            username: format!("__test_{}__", ObjectId::new().to_hex()),
-            password: TEST_PW.to_string(),
-        }
-    }
-
-    fn user(&self) -> &str {
-        &self.username
-    }
-
-    fn pw(&self) -> &str {
-        &self.password
-    }
-}
-
-impl Drop for TestUser {
-    fn drop(&mut self) {
-        let username = self.username.clone();
-        let handle = RT.handle().clone();
-        // Spawn a separate OS thread so we can block_on without nesting inside
-        // the RT.block_on() that is driving the test body.
-        std::thread::spawn(move || {
-            handle.block_on(async {
-                let _ = db::delete_user(username).await;
-            });
-        })
-        .join()
-        .ok();
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-fn parse_json<T: serde::de::DeserializeOwned>(body: &str) -> T {
-    serde_json::from_str(body).expect("Failed to parse JSON response")
-}
-
-/// Read the full response body as a String.
-async fn body_string(res: axum::response::Response) -> String {
-    let bytes = axum::body::to_bytes(res.into_body(), usize::MAX)
-        .await
-        .expect("Failed to read response body");
-    String::from_utf8(bytes.to_vec()).expect("Response body is not valid UTF-8")
-}
-
-/// Run an async test body on the shared runtime.
-/// Ensures the app (and DB) is initialized before entering the runtime.
-fn run<F: std::future::Future>(f: F) -> F::Output {
-    // Trigger app initialization BEFORE entering block_on, to avoid nested
-    // block_on calls (APP's LazyLock uses RT.block_on internally).
-    let _ = app();
-    RT.block_on(f)
-}
 
 // ===========================================================================
 // Tests
@@ -228,7 +96,7 @@ fn test_rate_limiting() {
     });
 }
 
-// ── Happy-path lifecycle ───────────────────────────────────────────────────
+// ── Happy-path lifecycle ─────────────────────────────────────────────────
 
 #[test]
 fn test_full_user_lifecycle() {
@@ -400,7 +268,7 @@ fn test_full_user_lifecycle() {
     });
 }
 
-// ── Error / edge-case tests ────────────────────────────────────────────────
+// ── Error / edge-case tests ──────────────────────────────────────────────
 
 #[test]
 fn test_duplicate_user_rejected() {
